requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coinmarketcap==3.0.1
idna==2.5
PyYAML==3.12
requests==2.18.3
urllib3==1.22
